pyproject.toml
[project]
name = "totp-tg-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "pyrofork>=2.3.66",
    "tgcrypto-pyrofork>=1.2.7",
    "tortoise-orm>=0.25.1",
]

[tool.pyright]
venvPath = '.'
venv = '.venv'
